+++ OUDS/Core/Themes/Orange/Tests/OrangeThemeTests.swift
@@ -96,6 +96,10 @@ struct OrangeThemeTests {
 
     // MARK: - Tests - Component tokens providers
 
+    @Test func defaultAccordionComponentTokensProvider() throws {
+        #expect(theme.accordion is OrangeThemeAccordionComponentTokensProvider)
+    }
+
     @Test func defaultAlertComponentTokensProvider() throws {
         #expect(theme.alert is OrangeThemeAlertComponentTokensProvider)
     }
@@ -140,6 +144,10 @@ struct OrangeThemeTests {
         #expect(theme.pinCodeInput is OrangeThemePinCodeInputComponentTokensProvider)
     }
 
+    @Test func defaultProgressIndicatorComponentTokensProvider() throws {
+        #expect(theme.progressIndicator is OrangeThemeProgressIndicatorComponentTokensProvider)
+    }
+
     @Test func defaultQuantityInputComponentTokensProvider() throws {
         #expect(theme.quantityInput is OrangeThemeQuantityInputComponentTokensProvider)
     }
@@ -174,6 +182,10 @@ struct OrangeThemeTests {
 
     @Test func defaultTextInputComponentTokensProvider() throws {
         #expect(theme.textInput is OrangeThemeTextInputComponentTokensProvider)
+    }
+
+    @Test func defaultTypographyComponentTokensProvider() throws {
+        #expect(theme.typography is OrangeThemeTypographyComponentTokensProvider)
     }
 }