+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -82,6 +82,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
         let sizes = SoshThemeSizeSemanticTokensProvider(dimensions: dimensions)
         let spaces = SoshThemeSpaceSemanticTokensProvider(dimensions: dimensions)
 
+        let accordion = SoshThemeAccordionComponentTokensProvider(sizes: sizes, spaces: spaces)
         let alert = SoshThemeAlertComponentTokensProvider(sizes: sizes, borders: borders, spaces: spaces)
         let badge = SoshThemeBadgeComponentTokensProvider(spaces: spaces, dimensions: dimensions)
         let bar = SoshThemeBarComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, opacities: opacities, effects: effects)
@@ -94,6 +95,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
         let link = SoshThemeLinkComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces)
         let listItem = SoshThemeListItemComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces, dimensions: dimensions)
         let pinCodeInput = SoshThemePinCodeInputComponentTokensProvider(spaces: spaces, dimensions: dimensions)
+        let progressIndicator = SoshThemeProgressIndicatorComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces)
         let quantityInput = SoshThemeQuantityInputComponentTokensProvider(sizes: sizes, spaces: spaces)
         let radioButton = SoshThemeRadioButtonComponentTokensProvider(sizes: sizes, borders: borders)
         let selectInput = SoshThemeSelectInputComponentTokensProvider(dimensions: dimensions)
@@ -103,6 +105,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
         let inputTag = SoshThemeInputTagComponentTokensProvider(borders: borders, colors: colors)
         let textArea = SoshThemeTextAreaComponentTokensProvider(sizes: sizes, spaces: spaces)
         let textInput = SoshThemeTextInputComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces, dimensions: dimensions)
+        let typography = SoshThemeTypographyComponentTokensProvider(spaces: spaces)
 
         super.init(borders: borders,
                    colors: colors,
@@ -115,6 +118,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
                    dimensions: dimensions,
                    sizes: sizes,
                    spaces: spaces,
+                   accordion: accordion,
                    alert: alert,
                    badge: badge,
                    bar: bar,
@@ -127,6 +131,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
                    link: link,
                    listItem: listItem,
                    pinCodeInput: pinCodeInput,
+                   progressIndicator: progressIndicator,
                    quantityInput: quantityInput,
                    radioButton: radioButton,
                    selectInput: selectInput,
@@ -136,6 +141,7 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
                    inputTag: inputTag,
                    textArea: textArea,
                    textInput: textInput,
+                   typography: typography,
                    resourcesBundle: Bundle.SoshTheme,
                    name: Self.name,
                    fontFamily: SoshBrandFontRawTokens.familyDefault,