+++ OUDS/Core/Themes/Sosh/Tests/SoshThemeTests.swift
@@ -98,6 +98,10 @@ struct SoshThemeTests {
 
     // MARK: - Tests - Component tokens providers
 
+    @Test func defaultAccordionComponentTokensProvider() throws {
+        #expect(theme.accordion is SoshThemeAccordionComponentTokensProvider)
+    }
+
     @Test func defaultAlertComponentTokensProvider() throws {
         #expect(theme.alert is SoshThemeAlertComponentTokensProvider)
     }
@@ -142,6 +146,10 @@ struct SoshThemeTests {
         #expect(theme.pinCodeInput is SoshThemePinCodeInputComponentTokensProvider)
     }
 
+    @Test func defaultProgressIndicatorComponentTokensProvider() throws {
+        #expect(theme.progressIndicator is SoshThemeProgressIndicatorComponentTokensProvider)
+    }
+
     @Test func defaultQuantityInputComponentTokensProvider() throws {
         #expect(theme.quantityInput is SoshThemeQuantityInputComponentTokensProvider)
     }
@@ -176,6 +184,10 @@ struct SoshThemeTests {
 
     @Test func defaultTextInputComponentTokensProvider() throws {
         #expect(theme.textInput is SoshThemeTextInputComponentTokensProvider)
+    }
+
+    @Test func defaultTypographyComponentTokensProvider() throws {
+        #expect(theme.typography is SoshThemeTypographyComponentTokensProvider)
     }
 }